Refactor code to be more OOP based
